Udemy Verification Forwarding/requirements.txt
Flask==3.0.3
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-dotenv==1.0.1
timeago==1.0.16
